Refractoring code
