+++ main.rb
@@ -1,22 +1,5 @@
 require './app'
 
-def options
-  puts ''
-  puts 'Welcome to School Library App!'
-  puts ''
-
-  puts '1 - List all books',
-       '2 - List all people',
-       '3 - Create a person',
-       '4 - Create a book',
-       '5 - Create a rental',
-       '6 - List all of rentals for a given person id',
-       '7 - Exit'
-  puts ''
-  puts 'Please choose an option by entering a number: '
-  print '#=> '
-end
-
 def select_option(app, num)
   case num
   when '1'
@@ -35,20 +18,6 @@ def select_option(app, num)
     puts ''
     puts '#=> Invalid option.(InvalidInputError)'
   end
-end
-
-def main
-  app = App.new
-  loop do
-    options
-    num = gets.chomp
-    break if num == '7'
-
-    select_option(app, num)
-    puts "\n"
-  end
-
-  puts 'Goodbye'
 end
 
 main
